Update login validation logic in forms.js
Changed the login form validation to only prevent submission and show an alert when both username and password are invalid, instead of on any invalid input. When correct credentials are provided, php code will create a session for the user that loggen in and allow is to show the orders for that particular user in the customer profile page.